Fix to.get is not a function.
When the event for updating the pagination is triggered on line `67` it tries to access a jQuery method `.get()` on the `to` object. The problem is that in my particular case the for some reason the `to` is not a jQuery object so it can't access the `.get()` method and it fails to update the pagination.

Just by wrapping the `to` to `$()` like this `$(to)` the error goes away.